what does a tui in scala look like?
